requirements.txt
africastalking==1.2.7
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.40.0
crispy-bootstrap4==2024.1
Django==5.0.4
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-filter==24.2
django-formtools==2.5.1
django-geojson==4.0.0
django-leaflet==0.29.1
django-location-field==2.7.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
fiona==1.9.6
frozenlist==1.4.1
geoip2==4.8.0
geojson==3.1.0
geopandas==0.14.4
idna==3.7
maxminddb==2.6.1
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
schema==0.7.7
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
